# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "HealthManager API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
  "paths": {
...
    "/products": {
      "post": {
        "operationId": "products_create",
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductRequest"
              }
            }
          },
          "required": true
        }
      },
      "get": {
        "operationId": "products_list",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "default": 1,
              "type": "integer"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "default": 20,
              "type": "integer"
            }
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedProductResponse"
                }
              }
            },
            "description": "OK"
          }
        }
      }
    },
    "/products/{id}": {
      "delete": {
        "operationId": "products_delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      },
      "put": {
        "operationId": "products_update",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductRequest"
              }
            }
          },
          "required": true
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ProductRequest": {
        "properties": {
          "price": {
            "format": "decimal",
            "type": "number",
            "minimum": 0.01
          },
          "applicationCount": {
            "type": "integer",
            "minimum": 1
          },
          "name": {
            "maxLength": 100,
            "type": "string"
          },
          "isActive": {
            "default": true,
            "type": "boolean"
          }
        },
        "type": "object",
        "required": [
          "name",
          "price",
          "applicationCount"
        ]
      },
      "ProductResponse": {
        "properties": {
          "price": {
            "type": "number",
            "format": "decimal"
          },
          "applicationCount": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "isActive": {
            "type": "boolean"
          }
        },
        "type": "object"
      },
      "PagedProductResponse": {
        "properties": {
          "total": {
            "type": "integer"
          },
          "pageSize": {
            "type": "integer"
          },
          "page": {
            "type": "integer"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ProductResponse"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
...
    }
  }
}